Remove _previous_name, copy extra config/headers:
Remove the _previous_name hack. It was added because a problem
appeared with an altered Addheader plugin that sent headers to
the backend service (instead of the frontend).

The headers would not go through because the new service had a
new name and the Addheader plugin was mapped to the original
service's name.

Passing the _previous_name allowed introspection in the plugin
to find the name of the original service.

However, Perlbal provides a "header" command, which does the same
thing as a reversed Addheader plugin. It doesn't map to service
names but instead uses the "extra_headers" key available in a
service. We simply ignored this key until now.

So now we copy over the extra_headers key and another one we
spotted: extra_config. Couldn't find another key worth copying.

Because of this there is no longer any need for _previous_name.
This is why it was removed. You now know.